mvp проекта

+++ src/schemas/rooms.py
@@ -6,7 +6,7 @@ from src.schemas.facilities import Facility
 class RoomRequestAdd(BaseModel):
     title: str = Field(min_length=1, max_length=100)
     description: str | None = Field(None)
-    price: int = Field(gt=0)
+    price: float = Field(gt=0)
     quantity: int = Field(gt=0)
     facilities: list[PositiveInt] = Field(default_factory=list)
 
@@ -15,7 +15,7 @@ class RoomAdd(BaseModel):
     hotel_id: PositiveInt
     title: str = Field(min_length=1, max_length=100)
     description: str | None = Field(None)
-    price: int = Field(gt=0)
+    price: float = Field(gt=0)
     quantity: int = Field(gt=0)
 
 
@@ -30,7 +30,7 @@ class RoomWithRels(Room):
 class RoomRequestPATCH(BaseModel):
     title: str | None = Field(None, min_length=1, max_length=100)
     description: str | None = Field(None)
-    price: int | None = Field(None, gt=0)
+    price: float | None = Field(None, gt=0)
     quantity: int | None = Field(None, gt=0)
     facilities: list[PositiveInt] = Field(default_factory=list)
 
@@ -38,5 +38,5 @@ class RoomRequestPATCH(BaseModel):
 class RoomPATCH(BaseModel):
     title: str | None = Field(None, min_length=1, max_length=100)
     description: str | None = Field(None)
-    price: int | None = Field(None, gt=0)
+    price: float | None = Field(None, gt=0)
     quantity: int | None = Field(None, gt=0)
